+++ src/js/Circle.js
@@ -188,7 +188,7 @@ class Circle {
 } // END class
 exports.Circle = Circle;
 Circle.circleUtils = {
-    vertAt: function (angle, radius) {
+    vertAt: (angle, radius) => {
         /* return new Vertex( Math.sin(angle) * radius,
                    Math.cos(angle) * radius ); */
         return new Vertex_1.Vertex(Math.cos(angle) * radius, Math.sin(angle) * radius);