+++ src/js/utils/datastructures/Color.d.ts
@@ -6,6 +6,8 @@
  * @modified 2019-11-18 Added a generic parse(string) function that detects the format.
  * @modified 2020-01-09 Fixed a bug in the parse(string) function. Hex colors with only three elements were considered faulty.
  * @modified 2020-10-23 Ported to Typescript.
+ * @modified 2021-02-08 Fixed a lot of es2015 compatibility issues.
+ * @version 0.0.9
  **/
 /**
  * @classdesc A color class, inspired by neolitec's Javascript class.
@@ -81,37 +83,40 @@ export declare class Color {
     fadein(v: string | number): void;
     fadeout(v: string | number): void;
     spin(v: string | number): void;
-    /** Debug */
     static makeRGB(...args: any[]): Color;
     static makeHSL(...args: Array<number | string>): Color;
     static makeHEX(value: string): Color;
+    /**
+     * Parse the given color string. Currently only these formate are recognized: hex, rgb, rgba.
+     *
+     * @method parse
+     * @static
+     * @memberof Color
+     * @param {string} str - The string representation to parse.
+     * @return {Color} The color instance that's represented by the given string.
+     */
     static parse(str: string): Color;
-    static Sanitizer: {
-        RGB: (...args: any[]) => any[];
-        HSL: (...args: Array<string | number>) => Array<number>;
-    };
-    static Validator: {
-        /**
-         * Check a hexa color (without #)
-         */
-        checkHEX: (value: any) => void;
-    };
-    static Convertor: {
-        /**
-         * Calculates HSL Color
-         * RGB must be normalized
-         * Must be executed in a Color object context
-         * http://mjijackson.com/2008/02/rgb-to-hsl-and-rgb-to-hsv-color-model-conversion-algorithms-in-javascript
-         */
-        RGBToHSL: () => void;
-        /**
-         * Calculates RGB color (nomalized)
-         * HSL must be normalized
-         * Must be executed in a Color object context
-         * http://mjijackson.com/2008/02/rgb-to-hsl-and-rgb-to-hsv-color-model-conversion-algorithms-in-javascript
-         */
-        HSLToRGB: () => void;
-    };
+    private static Sanitizer;
+    private static Validator;
+    private static Converter;
+    /**
+     * Create a clone of this color (RGB).
+     *
+     * @method clone
+     * @instance
+     * @memberof Color
+     * @return {Color} A clone of this color (in RGB mode).
+     */
     clone(): Color;
+    /**
+     * Interpolate this color on the RGB scale.
+     *
+     * @method interpolate
+     * @instance
+     * @memberof Color
+     * @param {Color} c - The color to interpolate to.
+     * @param {number} t - An interpolation value between 0.0 and 1.0.
+     * @return {Color} A clone of this color (in RGB mode).
+     */
     interpolate(c: Color, t: number): Color;
 }