+++ src/ts/utils/helpers/BezierPathInteractionHelper.ts
@@ -25,7 +25,10 @@
  * @public
  **/
 
-import { AlloyFinger } from "alloyfinger"; // node_modules
+// import { AlloyFinger } from "alloyfinger"; // node_modules
+// I would like to use AlloyFinger from the node_modules, but it seems
+// AlloyTeam has forgotten to publish their d.ts file.
+import { AlloyFinger } from "../../../../lib/alloy_finger";
 import { BezierPath } from "../../BezierPath";
 import { CubicBezierCurve } from "../../CubicBezierCurve";
 import { KeyHandler } from "../../KeyHandler";